re-vamped update R. No more bugs and pulling heavy dependencies!

+++ update-R.sh
@@ -65,43 +65,54 @@ function update_r {
 		remotemirror="http://cran.us.r-project.org"
 	fi
 	sudo chown -R ${USER} $(get_home_dir)/R
-	Rscript -e "update.packages(ask = FALSE, repos=\"${remotemirror}\")"
-	Rscript -e "update.packages(ask = FALSE, repos=\"${remotemirror}\")"
-	Rscript -e "if(!require(\"devtools\")) {install.packages(\"devtools\", ask=FALSE, repos=\"${remotemirror}\");devtools::install_github(\"hadley/devtools\")}"
-	Rscript -e 'if(!require("dtupdate")) devtools::install_github("hrbrmstr/dtupdate"); dtupdate::github_update()'
-	Rscript -e 'dtupdate::github_update(auto.install = TRUE, ask = FALSE)'
+	Rscript -e 'dir.create(path = Sys.getenv("R_LIBS_USER"), showWarnings = FALSE, recursive = TRUE)'
+	Rscript -e "update.packages(ask = FALSE, lib = Sys.getenv(\"R_LIBS_USER\"), repos=\"${remotemirror}\")"
 #	Rscript -e "update.packages(ask = FALSE, repos=\"${remotemirror}\")"
+#	Rscript -e "if(!require(\"devtools\")) {install.packages(\"devtools\", ask=FALSE, repos=\"${remotemirror}\");devtools::install_github(\"hadley/devtools\")}"
+#	Rscript -e 'if(!require("dtupdate")) devtools::install_github("hrbrmstr/dtupdate"); dtupdate::github_update()'
+#	Rscript -e 'dtupdate::github_update(auto.install = TRUE, ask = FALSE)'
+#	Rscript -e "update.packages(ask = FALSE, repos=\"${remotemirror}\")"
+}
+
+function get_netversion {
+#		netversion=$(Rscript -e 'cat(stringr::str_match(scan("https://www.rstudio.org/links/check_for_update?version=1.0.0", what = character(0), quiet=TRUE), "^[^=]+=([^\\&]+)\\&.*")[[2]])')
+#		netversion=$(wget --no-check-certificate -qO- https://s3.amazonaws.com/rstudio-server/current.ver)
+#   	netversion=$(curl -s https://www.rstudio.org/links/check_for_update?version=1.0.0 | grep -oEi 'update-version=([0-9]+\.[0-9]+\.[0-9]+)' | awk -F= '{print $2}')
+# 	netversion=$(wget --no-check-certificate -qO- https://s3.amazonaws.com/rstudio-server/current.ver)
+
+	netversion=$(curl -s http://download1.rstudio.org/current.ver)
+	pattern='^([0-9.+]+)\+([0-9]+)'
+	if [[ $netversion =~ $pattern ]]; then
+		netversion1=${BASH_REMATCH[1]}
+		netversion2=${BASH_REMATCH[2]}
+		netversion="${netversion1}+${netversion2}"
+	else
+		return
+	fi
 }
 
 function update_rstudio {
 	ver=$(apt show rstudio | grep Version)
-	pattern='^Version: ([0-9.]+)\s*$'
+	if [[ "$ver" == "" ]]; then
+		return
+	fi
+	pattern='^Version: ([0-9.+]+)\s*$'
 	if [[ $ver =~ $pattern ]]; then
 		ourversion=${BASH_REMATCH[1]}
-#		netversion=$(Rscript -e 'cat(stringr::str_match(scan("https://www.rstudio.org/links/check_for_update?version=1.0.0", what = character(0), quiet=TRUE), "^[^=]+=([^\\&]+)\\&.*")[[2]])')
-#		netversion=$(wget --no-check-certificate -qO- https://s3.amazonaws.com/rstudio-server/current.ver)
-		netversion=$(curl -s http://download1.rstudio.org/current.ver)
-		if [ "$ourversion" != "$netversion" ]; then
-			deb_folder=$(get_deb_folder)
-			tee /tmp/get_rstudio_uri.R <<EOF
-if(!require('rvest')) install.packages('rvest', Ncpus=8, repos='${remotemirror}')
-xpath='.downloads:nth-child(2) tr:nth-child(5) a'
-url = "https://www.rstudio.com/products/rstudio/download/"
-thepage<-xml2::read_html(url)
-cat(html_node(thepage, xpath) %>% html_attr("href"))
-EOF
-			RSTUDIO_URI=$(Rscript /tmp/get_rstudio_uri.R)
-			
-			wget -c --output-document /tmp/rstudio.deb $RSTUDIO_URI -O ${deb_folder}/rstudio_${netversion}_amd64.deb
-			if ! sudo dpkg -i ${deb_folder}/rstudio_${netversion}_amd64.deb; then
-				sudo apt install -f --yes
-			fi
-			rm /tmp/get_rstudio_uri.R
-	
-			if fc-list |grep -q FiraCode; then
-				if ! grep -q "text-rendering:" /usr/lib/rstudio/www/index.htm; then
-					sudo sed -i '/<head>/a<style>*{text-rendering: optimizeLegibility;}<\/style>' /usr/lib/rstudio/www/index.htm
-				fi
+  	fi
+	if [ "$ourversion" != "$netversion" ]; then
+		deb_folder=$(get_deb_folder)
+		RSTUDIO_URI="https://download1.rstudio.org/desktop/bionic/amd64/rstudio-${netversion1}-${netversion2}-amd64.deb"
+		
+		wget -c --output-document /tmp/rstudio.deb $RSTUDIO_URI -O ${deb_folder}/rstudio_${netversion}_amd64.deb
+		if ! sudo dpkg -i ${deb_folder}/rstudio_${netversion}_amd64.deb; then
+			sudo apt install -f --yes
+		fi
+		rm /tmp/get_rstudio_uri.R
+
+		if fc-list |grep -q FiraCode; then
+			if ! grep -q "text-rendering:" /usr/lib/rstudio/www/index.htm; then
+				sudo sed -i '/<head>/a<style>*{text-rendering: optimizeLegibility;}<\/style>' /usr/lib/rstudio/www/index.htm
 			fi
 		fi
 	fi
@@ -109,25 +120,17 @@ EOF
 
 function update_rstudio_server {
 	ver=$(apt show rstudio-server | grep Version)
-	pattern='^Version: ([0-9.]+)\s*$'
+	if [[ "$ver" == "" ]]; then
+		return
+	fi
+	pattern='^Version: ([0-9.+]+)\s*$'
 	if [[ $ver =~ $pattern ]]; then
 		ourversion=${BASH_REMATCH[1]}
 		do_check=1
 	fi
-	netversion=$(wget --no-check-certificate -qO- https://s3.amazonaws.com/rstudio-server/current.ver)
 	if [ "$ourversion" != "$netversion" ]; then
 		deb_folder=$(get_deb_folder)
-		tee /tmp/get_rstudio_uri.R <<'EOT'
-if(!require('rvest')) install.packages('rvest', Ncpus=8, repos='http://cran.us.r-project.org')
-if(!require('stringr')) install.packages('stringr', Ncpus=8, repos='http://cran.us.r-project.org')
-xpath='code:nth-child(3)'
-url = "https://www.rstudio.com/products/rstudio/download-server/"
-thepage<-xml2::read_html(url)
-link<-html_node(thepage, xpath) %>% html_text()
-cat(stringr::str_match(link, '^\\$( wget)? (.*)$')[[3]])
-EOT
-		RSTUDIO_URI=$(Rscript /tmp/get_rstudio_uri.R)
-		RSTUDIO_URI=$(Rscript /tmp/get_rstudio_uri.R)
+		RSTUDIO_URI="wget https://download2.rstudio.org/server/bionic/amd64/rstudio-server-${netversion1}-${netversion2}-amd64.deb"
 		
 		wget -c $RSTUDIO_URI --output-document ${deb_folder}/rstudio-server_${netversion}_amd64.deb
 		if ! sudo dpkg -i ${deb_folder}/rstudio-server_${netversion}_amd64.deb; then
@@ -141,11 +144,9 @@ if dpkg -s r-base-core >/dev/null 2>/dev/null; then
 	update_r
 fi
 
-if dpkg -s rstudio >/dev/null 2>/dev/null; then
-	update_rstudio
-fi
+get_netversion
 
-if dpkg -s rstudio-server >/dev/null 2>/dev/null; then
-	update_rstudio_server
-fi
+update_rstudio
+
+update_rstudio_server
